+++ src/index.ts
@@ -1,1 +1,7 @@
+import { polyfillWebCrypto } from './polyfillWebCrypto';
+
 export * from './NativeCryptoModule';
+export * from './getRandomValues';
+export * from './polyfillWebCrypto';
+
+polyfillWebCrypto();